+++ packages/xrpl/src/models/common/index.ts
@@ -217,27 +217,3 @@ export interface PriceData {
     Scale?: number
   }
 }
-
-/**
- * MPTokenMetadata object as per the XLS-89d standard.
- */
-export interface MPTokenMetadata {
-  ticker: string
-  name: string
-  desc: string
-  icon: string
-  asset_class: string
-  asset_subclass?: string
-  issuer_name: string
-  urls?: MPTokenMetadataUrl[]
-  additional_info?: string
-}
-
-/**
- * MPTokenMetadataUrl object as per the XLS-89d standard.
- */
-export interface MPTokenMetadataUrl {
-  url: string
-  type: string
-  title: string
-}
