+++ packages/xrpl/test/fixtures/transactions/mptokenMetadata.json
@@ -1,0 +1,217 @@
+[
+  {
+    "testName": "valid MPTokenMetadata",
+    "mptMetadata": {
+      "ticker": "TBILL",
+      "name": "T-Bill Yield Token",
+      "desc": "A yield-bearing stablecoin backed by short-term U.S. Treasuries and money market instruments.",
+      "icon": "https://example.org/tbill-icon.png",
+      "asset_class": "rwa",
+      "asset_subclass": "treasury",
+      "issuer_name": "Example Yield Co.",
+      "urls": [
+        {
+          "url": "https://exampleyield.co/tbill",
+          "type": "website",
+          "title": "Product Page"
+        },
+        {
+          "url": "https://exampleyield.co/docs",
+          "type": "docs",
+          "title": "Yield Token Docs"
+        }
+      ],
+      "additional_info": {
+        "interest_rate": "5.00%",
+        "interest_type": "variable",
+        "yield_source": "U.S. Treasury Bills",
+        "maturity_date": "2045-06-30",
+        "cusip": "912796RX0"
+      }
+    },
+    "warningsMessages": []
+  },
+  {
+    "testName": "missing ticker",
+    "mptMetadata": {
+      "name": "T-Bill Yield Token",
+      "desc": "A yield-bearing stablecoin backed by short-term U.S. Treasuries and money market instruments.",
+      "icon": "https://example.org/tbill-icon.png",
+      "asset_class": "rwa",
+      "asset_subclass": "treasury",
+      "issuer_name": "Example Yield Co.",
+      "urls": [
+        {
+          "url": "https://exampleyield.co/tbill",
+          "type": "website",
+          "title": "Product Page"
+        },
+        {
+          "url": "https://exampleyield.co/docs",
+          "type": "docs",
+          "title": "Yield Token Docs"
+        }
+      ],
+      "additional_info": {
+        "interest_rate": "5.00%",
+        "interest_type": "variable",
+        "yield_source": "U.S. Treasury Bills",
+        "maturity_date": "2045-06-30",
+        "cusip": "912796RX0"
+      }
+    },
+    "warningsMessages": [
+      "ticker is required and must be string."
+    ]
+  },
+  {
+    "testName": "ticker has lowercase letters",
+    "mptMetadata": {
+      "ticker": "tbill",
+      "name": "T-Bill Token",
+      "icon": "https://example.com/icon.png",
+      "asset_class": "rwa",
+      "asset_subclass": "treasury",
+      "issuer_name": "Issuer"
+    },
+    "warningsMessages": [
+      "ticker should have uppercase letters (A-Z) and digits (0-9) only. Max 6 characters recommended."
+    ]
+  },
+  {
+    "testName": "icon not https",
+    "mptMetadata": {
+      "ticker": "TBILL",
+      "name": "T-Bill Token",
+      "icon": "http://example.com/icon.png",
+      "asset_class": "rwa",
+      "asset_subclass": "treasury",
+      "issuer_name": "Issuer"
+    },
+    "warningsMessages": [
+      "icon should be a valid https url."
+    ]
+  },
+  {
+    "testName": "invalid asset_class",
+    "mptMetadata": {
+      "ticker": "TBILL",
+      "name": "T-Bill Token",
+      "icon": "https://example.com/icon.png",
+      "asset_class": "invalid",
+      "asset_subclass": "treasury",
+      "issuer_name": "Issuer"
+    },
+    "warningsMessages": [
+      "asset_class should be one of rwa, memes, wrapped, gaming, defi, other."
+    ]
+  },
+  {
+    "testName": "invalid asset_subclass",
+    "mptMetadata": {
+      "ticker": "TBILL",
+      "name": "T-Bill Token",
+      "icon": "https://example.com/icon.png",
+      "asset_class": "rwa",
+      "asset_subclass": "junk",
+      "issuer_name": "Issuer"
+    },
+    "warningsMessages": [
+      "asset_subclass should be one of stablecoin, commodity, real_estate, private_credit, equity, treasury, other."
+    ]
+  },
+  {
+    "testName": "missing asset_subclass for rwa",
+    "mptMetadata": {
+      "ticker": "TBILL",
+      "name": "T-Bill Token",
+      "icon": "https://example.com/icon.png",
+      "asset_class": "rwa",
+      "issuer_name": "Issuer"
+    },
+    "warningsMessages": [
+      "asset_subclass is required when asset_class is rwa."
+    ]
+  },
+  {
+    "testName": "urls not https",
+    "mptMetadata": {
+      "ticker": "TBILL",
+      "name": "T-Bill Token",
+      "icon": "https://example.com/icon.png",
+      "asset_class": "defi",
+      "issuer_name": "Issuer",
+      "asset_subclass": "stablecoin",
+      "urls": [
+        {
+          "url": "http://notsecure.com",
+          "type": "website",
+          "title": "Homepage"
+        }
+      ]
+    },
+    "warningsMessages": [
+      "url should be a valid https url."
+    ]
+  },
+  {
+    "testName": "additional_info is invalid type",
+    "mptMetadata": {
+      "ticker": "TBILL",
+      "name": "T-Bill Token",
+      "icon": "https://example.com/icon.png",
+      "asset_class": "defi",
+      "issuer_name": "Issuer",
+      "additional_info": [
+        "not",
+        "valid"
+      ]
+    },
+    "warningsMessages": [
+      "additional_info must be a string or JSON object."
+    ]
+  },
+  {
+    "testName": "multiple warnings",
+    "mptMetadata": {
+      "ticker": "TBILLLLLLL",
+      "name": "T-Bill Yield Token",
+      "desc": "A yield-bearing stablecoin backed by short-term U.S. Treasuries and money market instruments.",
+      "icon": "https/example.org/tbill-icon.png",
+      "asset_class": "rwamemes",
+      "asset_subclass": "treasurymemes",
+      "issuer_name": "Example Yield Co.",
+      "urls": [
+        {
+          "url": "http://notsecure.com",
+          "type": "website",
+          "title": "Homepage"
+        }
+      ],
+      "additional_info": {
+        "interest_rate": "5.00%",
+        "interest_type": "variable",
+        "yield_source": "U.S. Treasury Bills",
+        "maturity_date": "2045-06-30",
+        "cusip": "912796RX0"
+      }
+    },
+    "warningsMessages": [
+      "ticker should have uppercase letters (A-Z) and digits (0-9) only. Max 6 characters recommended.",
+      "icon should be a valid https url.",
+      "asset_class should be one of rwa, memes, wrapped, gaming, defi, other.",
+      "asset_subclass should be one of stablecoin, commodity, real_estate, private_credit, equity, treasury, other.",
+      "url should be a valid https url."
+    ]
+  },
+  {
+    "testName": "null mptMetadata",
+    "mptMetadata": null,
+    "warningsMessages": []
+  },
+  {
+    "testName": "empty mptMetadata",
+    "mptMetadata": {},
+    "warningsMessages": []
+  }
+]
